Rename stores for clarity
