+++ src/ui/UserPortal/app.vue
@@ -11,7 +11,7 @@
 
 <script lang="ts">
 import { mapStores } from 'pinia';
-import { appConfig } from '@/stores/appConfig';
+import { appConfig } from '@/stores/appConfigStore';
 
 export default {
 	data() {
