+++ src/ui/UserPortal/components/NavBar.vue
@@ -64,8 +64,8 @@
 <script lang="ts">
 import { mapStores } from 'pinia';
 import type { Session } from '@/js/types';
-import { appConfig } from '@/stores/appConfig';
-import { appStore } from '@/stores/app';
+import { appConfig } from '@/stores/appConfigStore';
+import { appStore } from '@/stores/appStore';
 import { getMsalInstance, getLoginRequest } from '@/js/auth';
 
 export default {
